refactor: epochs routes
