--- src/routes/epochs/index.ts
@@ -1,486 +1,0 @@
-import { FastifyInstance, FastifyRequest } from 'fastify';
-import * as QueryTypes from '../../types/queries/epochs';
-import * as ResponseTypes from '../../types/responses/epochs';
-import { getDbSync } from '../../utils/database';
-import { getSchemaForEndpoint } from '@blockfrost/openapi';
-import { handle400Custom, handle404 } from '../../utils/error-handler';
-import { validateAndConvertPool, validatePositiveInRangeSignedInt } from '../../utils/validation';
-import { SQLQuery } from '../../sql';
-import { sortKeysInObject } from '../../utils/string-utils';
-import { GENESIS } from '../../constants/genesis';
-import { getConfig } from '../../config';
-
-async function epochs(fastify: FastifyInstance) {
-  fastify.route({
-    url: '/epochs/latest',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/latest'),
-    handler: async (_request, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        const network = getConfig().network;
-        const epochLength = GENESIS[network].epoch_length;
-        const { rows }: { rows: ResponseTypes.Epoch[] } =
-          await clientDbSync.query<QueryTypes.Epoch>(SQLQuery.get('epochs_latest'), [epochLength]);
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return handle404(reply);
-        }
-
-        return reply.send(rows[0]);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-
-  fastify.route({
-    url: '/epochs/:number',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/{number}'),
-    handler: async (request: FastifyRequest<QueryTypes.RequestParameters>, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        if (!validatePositiveInRangeSignedInt(request.params.number)) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Missing, out of range or malformed epoch_number.');
-        }
-
-        const network = getConfig().network;
-        const epochLength = GENESIS[network].epoch_length;
-        const { rows }: { rows: ResponseTypes.Epoch[] } =
-          await clientDbSync.query<QueryTypes.Epoch>(SQLQuery.get('epochs_number'), [
-            request.params.number,
-            epochLength,
-          ]);
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return handle404(reply);
-        }
-
-        return reply.send(rows[0]);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-
-  fastify.route({
-    url: '/epochs/:number/next',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/{number}/next'),
-    handler: async (request: FastifyRequest<QueryTypes.RequestParameters>, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        if (!validatePositiveInRangeSignedInt(request.params.number)) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Missing, out of range or malformed epoch_number.');
-        }
-
-        const query404 = await clientDbSync.query<QueryTypes.ResultFound>(
-          SQLQuery.get('epochs_404'),
-          [request.params.number],
-        );
-
-        if (query404.rows.length === 0) {
-          clientDbSync.release();
-          return handle404(reply);
-        }
-
-        const network = getConfig().network;
-        const epochLength = GENESIS[network].epoch_length;
-
-        const { rows }: { rows: ResponseTypes.Epoch[] } =
-          await clientDbSync.query<QueryTypes.Epoch>(SQLQuery.get('epochs_number_next'), [
-            request.params.number,
-            request.query.count,
-            request.query.page,
-            epochLength,
-          ]);
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return reply.send([]);
-        }
-
-        return reply.send(rows);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-
-  fastify.route({
-    url: '/epochs/:number/previous',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/{number}/previous'),
-    handler: async (request: FastifyRequest<QueryTypes.RequestParameters>, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        if (!validatePositiveInRangeSignedInt(request.params.number)) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Missing, out of range or malformed epoch_number.');
-        }
-
-        const query404 = await clientDbSync.query<QueryTypes.ResultFound>(
-          SQLQuery.get('epochs_404'),
-          [request.params.number],
-        );
-
-        if (query404.rows.length === 0) {
-          clientDbSync.release();
-          return handle404(reply);
-        }
-
-        const network = getConfig().network;
-        const epochLength = GENESIS[network].epoch_length;
-
-        const { rows }: { rows: ResponseTypes.Epoch[] } =
-          await clientDbSync.query<QueryTypes.Epoch>(SQLQuery.get('epochs_number_previous'), [
-            request.params.number,
-            request.query.count,
-            request.query.page,
-            epochLength,
-          ]);
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return reply.send([]);
-        }
-
-        return reply.send(rows);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-
-  fastify.route({
-    url: '/epochs/:number/stakes',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/{number}/stakes'),
-    handler: async (request: FastifyRequest<QueryTypes.RequestStakeParameters>, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        if (!validatePositiveInRangeSignedInt(request.params.number)) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Missing, out of range or malformed epoch_number.');
-        }
-
-        const query404 = await clientDbSync.query<QueryTypes.ResultFound>(
-          SQLQuery.get('epochs_404'),
-          [request.params.number],
-        );
-
-        if (query404.rows.length === 0) {
-          clientDbSync.release();
-          return handle404(reply);
-        }
-
-        const { rows }: { rows: ResponseTypes.EpochStakes } =
-          await clientDbSync.query<QueryTypes.EpochStake>(SQLQuery.get('epochs_number_stakes'), [
-            request.params.number,
-            request.query.count,
-            request.query.page,
-          ]);
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return reply.send([]);
-        }
-
-        return reply.send(rows);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-
-  fastify.route({
-    url: '/epochs/:number/stakes/:pool_id',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/{number}/stakes/{pool_id}'),
-    handler: async (request: FastifyRequest<QueryTypes.RequestStakePoolIdParameters>, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        if (!validatePositiveInRangeSignedInt(request.params.number)) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Missing, out of range or malformed epoch_number.');
-        }
-
-        const query404_epoch = await clientDbSync.query<QueryTypes.ResultFound>(
-          SQLQuery.get('epochs_404'),
-          [request.params.number],
-        );
-
-        if (query404_epoch.rows.length === 0) {
-          clientDbSync.release();
-          return handle404(reply);
-        }
-
-        // validate (and convert hex->bech32 if needed) pool ID
-        const pool_id = validateAndConvertPool(request.params.pool_id);
-
-        if (!pool_id) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Invalid or malformed pool id format.');
-        }
-
-        const query404_pool = await clientDbSync.query<QueryTypes.ResultFound>(
-          SQLQuery.get('epochs_pool_404'),
-          [pool_id],
-        );
-
-        if (query404_pool.rows.length === 0) {
-          clientDbSync.release();
-          return handle404(reply);
-        }
-
-        const { rows }: { rows: ResponseTypes.EpochStakesPoolId } =
-          await clientDbSync.query<QueryTypes.EpochStakesPoolId>(
-            SQLQuery.get('epochs_number_stakes_pool_id'),
-            [request.params.number, request.query.count, request.query.page, pool_id],
-          );
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return reply.send([]);
-        }
-
-        return reply.send(rows);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-
-  fastify.route({
-    url: '/epochs/:number/blocks',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/{number}/blocks'),
-    handler: async (request: FastifyRequest<QueryTypes.RequestBlockParameters>, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        if (!validatePositiveInRangeSignedInt(request.params.number)) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Missing, out of range or malformed epoch_number.');
-        }
-
-        const query404 = await clientDbSync.query<QueryTypes.ResultFound>(
-          SQLQuery.get('epochs_404'),
-          [request.params.number],
-        );
-
-        if (query404.rows.length === 0) {
-          clientDbSync.release();
-          return handle404(reply);
-        }
-
-        const { rows } = await clientDbSync.query<QueryTypes.EpochBlocks>(
-          SQLQuery.get('epochs_number_blocks'),
-          [request.query.order, request.query.count, request.query.page, request.params.number],
-        );
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return reply.send([]);
-        }
-
-        const list: string[] = [];
-
-        for (const row of rows) {
-          list.push(row.hash);
-        }
-
-        return reply.send(list);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-
-  fastify.route({
-    url: '/epochs/:number/blocks/:pool_id',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/{number}/blocks/{pool_id}'),
-    handler: async (request: FastifyRequest<QueryTypes.RequestBlockPoolIdParameters>, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        if (!validatePositiveInRangeSignedInt(request.params.number)) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Missing, out of range or malformed epoch_number.');
-        }
-
-        const query404_epoch = await clientDbSync.query<QueryTypes.ResultFound>(
-          SQLQuery.get('epochs_404'),
-          [request.params.number],
-        );
-
-        if (query404_epoch.rows.length === 0) {
-          clientDbSync.release();
-          return handle404(reply);
-        }
-
-        // validate (and convert hex->bech32 if needed) pool ID
-        const pool_id = validateAndConvertPool(request.params.pool_id);
-
-        if (!pool_id) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Invalid or malformed pool id format.');
-        }
-
-        const query404_pool = await clientDbSync.query<QueryTypes.ResultFound>(
-          SQLQuery.get('epochs_pool_404'),
-          [pool_id],
-        );
-
-        if (query404_pool.rows.length === 0) {
-          clientDbSync.release();
-          return handle404(reply);
-        }
-
-        const { rows } = await clientDbSync.query<QueryTypes.EpochBlocksPoolId>(
-          SQLQuery.get('epochs_number_blocks_pool_id'),
-          [
-            request.query.order,
-            request.query.count,
-            request.query.page,
-            request.params.number,
-            pool_id,
-          ],
-        );
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return reply.send([]);
-        }
-
-        const list: string[] = [];
-
-        for (const row of rows) {
-          list.push(row.hash);
-        }
-
-        return reply.send(list);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-
-  fastify.route({
-    url: '/epochs/:number/parameters',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/{number}/parameters'),
-    handler: async (request: FastifyRequest<QueryTypes.RequestParameters>, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        if (!validatePositiveInRangeSignedInt(request.params.number)) {
-          clientDbSync.release();
-          return handle400Custom(reply, 'Missing, out of range or malformed epoch_number.');
-        }
-
-        const { rows }: { rows: ResponseTypes.EpochParameters[] } =
-          await clientDbSync.query<QueryTypes.EpochParameters>(
-            SQLQuery.get('epochs_number_parameters'),
-            [request.params.number],
-          );
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return handle404(reply);
-        }
-
-        if (rows[0].cost_models) {
-          rows[0].cost_models = sortKeysInObject(rows[0].cost_models);
-        }
-
-        return reply.send(rows[0]);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-
-  fastify.route({
-    url: '/epochs/latest/parameters',
-    method: 'GET',
-    schema: getSchemaForEndpoint('/epochs/latest/parameters'),
-    handler: async (_request, reply) => {
-      const clientDbSync = await getDbSync(fastify);
-
-      try {
-        const { rows }: { rows: ResponseTypes.EpochParameters[] } =
-          await clientDbSync.query<QueryTypes.EpochParameters>(
-            SQLQuery.get('epochs_latest_parameters'),
-          );
-
-        clientDbSync.release();
-
-        if (rows.length === 0) {
-          return handle404(reply);
-        }
-
-        if (rows[0].cost_models) {
-          rows[0].cost_models = sortKeysInObject(rows[0].cost_models);
-        }
-
-        return reply.send(rows[0]);
-      } catch (error) {
-        if (clientDbSync) {
-          clientDbSync.release();
-        }
-        throw error;
-      }
-    },
-  });
-}
-
-module.exports = epochs;
